refactor: 메인 클레스와 controller 분리
