--- src/main/java/chess/SparkChessApplication.java
@@ -1,16 +1,0 @@
-package chess;
-
-import chess.web.controller.ChessWebController;
-import chess.web.service.ChessService;
-
-import static spark.Spark.port;
-import static spark.Spark.staticFileLocation;
-
-public class SparkChessApplication {
-    public static void main(String[] args) {
-        port(8080);
-        staticFileLocation("static");
-        ChessWebController chessController = new ChessWebController(new ChessService());
-        chessController.run();
-    }
-}
